fix(backend-mavericks: Refactor the application user test
