+++ Backend/starter_project_mavericks/BlogApp/BlogApp.Application/Features/Authentication/CQRS/Handlers/SigninHandler.cs
@@ -1,8 +1,9 @@
 ﻿using BlogApp.Application.Contracts.Identity;
 using BlogApp.Application.Features.Authentication.CQRS.Commands;
 using BlogApp.Application.Features.Authentication.DTO;
-using BlogApp.Application.Features.Authentication.DTO.Validators;
 using BlogApp.Application.Responses;
+using BlogApp.Application.Features.Authentication.DTO.Validators;
+
 using MediatR;
 using System;
 using System.Collections.Generic;
